refactor: split ai_summary_card.dart
